fixed some major issues with frame timing fixed cfr sync issue fixed some major sync issues that may have been present in major version fixed a bug with audio output timing made mic data output more accurately as long as the qpc timestamp checks out

+++ Source/Settings.cpp
@@ -2178,7 +2178,7 @@ void OBS::ApplySettings()
 
                 //--------------------------------------------------
 
-                int globalAudioTimeAdjust = SendMessage(GetDlgItem(hwndCurrentSettings, IDC_AUDIOTIMEADJUST), UDM_GETPOS32, 0, 0);
+                int globalAudioTimeAdjust = (int)SendMessage(GetDlgItem(hwndCurrentSettings, IDC_AUDIOTIMEADJUST), UDM_GETPOS32, 0, 0);
                 GlobalConfig->SetInt(TEXT("Audio"), TEXT("GlobalAudioTimeAdjust"), globalAudioTimeAdjust);
 
                 //--------------------------------------------------
